refactor: split fixtures into multiple files
