+++ tests/apps/workshops/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 
-from quickwrench_api.apps.car_makes.models import CarMake
 from quickwrench_api.apps.accounts.models import Account
+from quickwrench_api.apps.car_makes.models import CarMake
 from quickwrench_api.apps.workshops.models import Category, Service, Workshop
